get timeline working

+++ services/tweet/internal/usecase/delete_tweet_usecase.go
@@ -2,8 +2,11 @@ package usecase
 
 import (
 	"context"
+	"fmt"
 
+	"github.com/Drivello/Twittah/services/tweet/internal/common"
 	"github.com/Drivello/Twittah/services/tweet/internal/ports"
+	"go.uber.org/zap"
 )
 
 type DeleteTweetUsecase struct {
@@ -14,6 +17,11 @@ func NewDeleteTweetUsecase(repo ports.TweetRepository) *DeleteTweetUsecase {
 	return &DeleteTweetUsecase{repo: repo}
 }
 
-func (uc *DeleteTweetUsecase) DeleteTweet(ctx context.Context, tweetID int64) error {
+func (uc *DeleteTweetUsecase) Execute(ctx context.Context, tweetID int64) error {
+	common.Logger().Debug("[DeleteTweetUsecase] Execute called", zap.Int64("tweet_id", tweetID))
+	if tweetID <= 0 {
+		return fmt.Errorf("tweet id must be greater than zero")
+	}
+
 	return uc.repo.Delete(tweetID)
 }
